delete a message now availeble

+++ src/main/java/ch/unibe/ese/team6/controller/service/MessageService.java
@@ -181,6 +181,10 @@ public class MessageService {
 	}
 
 	public void deleteMessage(long id) {
-		messageDao.delete(id);	
+		Message message = messageDao.findOne(id);
+		if(message != null) {
+			message.setState(MessageState.READ);
+			messageDao.delete(id);	
+		}	
 	}
 }
